backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.12
celery==5.6.3
redis==7.4.0
python-dotenv==1.2.2
requests==2.34.0
python-multipart==0.0.28
dnspython==2.8.0
slowapi==0.1.9
limits==5.8.0
python-json-logger==4.1.0
alembic==1.16.1
websockets==14.2
